# Conflicts:
#	src/python/ccpn/core/Peak.py

+++ src/python/ccpn/util/Update.py
@@ -5,8 +5,8 @@ Module Documentation here
 # Licence, Reference and Credits
 #=========================================================================================
 __copyright__ = "Copyright (C) CCPN project (https://www.ccpn.ac.uk) 2014 - 2023"
-__credits__ = ("Ed Brooksbank, Joanna Fox, Victoria A Higman, Luca Mureddu, Eliza Płoskoń",
-               "Timothy J Ragan, Brian O Smith, Gary S Thompson & Geerten W Vuister")
+__credits__ = ("Ed Brooksbank, Joanna Fox, Morgan Hayward, Victoria A Higman, Luca Mureddu",
+               "Eliza Płoskoń, Timothy J Ragan, Brian O Smith, Gary S Thompson & Geerten W Vuister")
 __licence__ = ("CCPN licence. See https://ccpn.ac.uk/software/licensing/")
 __reference__ = ("Skinner, S.P., Fogh, R.H., Boucher, W., Ragan, T.J., Mureddu, L.G., & Vuister, G.W.",
                  "CcpNmr AnalysisAssign: a flexible platform for integrated NMR analysis",
@@ -15,8 +15,8 @@ __reference__ = ("Skinner, S.P., Fogh, R.H., Boucher, W., Ragan, T.J., Mureddu, 
 # Last code modification
 #=========================================================================================
 __modifiedBy__ = "$modifiedBy: Ed Brooksbank $"
-__dateModified__ = "$dateModified: 2023-10-12 16:29:14 +0100 (Thu, October 12, 2023) $"
-__version__ = "$Revision: 3.2.0.1 $"
+__dateModified__ = "$dateModified: 2023-12-05 14:43:22 +0000 (Tue, December 05, 2023) $"
+__version__ = "$Revision: 3.2.1 $"
 #=========================================================================================
 # Created
 #=========================================================================================
@@ -303,10 +303,12 @@ class UpdateFile:
                         _name = _file.basename
 
                         # is a python file - remove the pycache to make sure that it loads correctly next time
-                        for fp in [Path.aPath(fp) for fp in (_file.filepath / '__pycache__').listdir()]:
-                            if fp.basename == _name and fp.suffix == '.pyc':
-                                # print(f'cleaning pycache: {fp.name}')
-                                fp.removeFile()
+                        pyDir = _file.filepath / '__pycache__'
+                        if pyDir.exists() and pyDir.is_dir():
+                            for fp in [Path.aPath(fp) for fp in pyDir.listdir()]:
+                                if fp.basename == _name and fp.suffix == '.pyc':
+                                    # print(f'cleaning pycache: {fp.name}')
+                                    fp.removeFile()
 
                     # generate the hashcode for the new file here
                     currentHashCode = calcHashCode(fullFilePath)
